introduce get-bot-credentials module
