+++ .devcontainer/devcontainer.json
@@ -30,6 +30,7 @@
 		"streetsidesoftware.code-spell-checker",
 		"vscode-icons-team.vscode-icons",
 		"hediet.vscode-drawio",
+		"GitHub.copilot"
 	],
 	"postCreateCommand": "yarn install",
 }
